minor code shuffling, restore newer version of testify packages
