+++ pkg/scheduler/pod_task.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	"code.google.com/p/go-uuid/uuid"
-	"code.google.com/p/goprotobuf/proto"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/api"
 	log "github.com/golang/glog"
 	"github.com/mesos/mesos-go/mesos"
@@ -24,28 +23,6 @@ type PodTask struct {
 	launched bool
 	deleted  bool
 	podKey   string
-}
-
-func rangeResource(name string, ports []uint64) *mesos.Resource {
-	if len(ports) == 0 {
-		// pod may consist of a container that doesn't expose any ports on the host
-		return nil
-	}
-	return &mesos.Resource{
-		Name:   proto.String(name),
-		Type:   mesos.Value_RANGES.Enum(),
-		Ranges: NewRanges(ports),
-	}
-}
-
-// func NewRange(begin uint64, end uint64) *mesos.Value_Ranges {
-func NewRanges(ports []uint64) *mesos.Value_Ranges {
-	r := make([]*mesos.Value_Range, 0)
-	for _, port := range ports {
-		x := proto.Uint64(port)
-		r = append(r, &mesos.Value_Range{Begin: x, End: x})
-	}
-	return &mesos.Value_Ranges{Range: r}
 }
 
 func (t *PodTask) hasAcceptedOffer() bool {
@@ -75,7 +52,7 @@ func (t *PodTask) FillTaskInfo(offer PerishableOffer) error {
 	t.Offer = offer
 	log.V(3).Infof("Recording offer(s) %v against pod %v", details.Id, t.Pod.Name)
 
-	t.TaskInfo.TaskId = &mesos.TaskID{Value: proto.String(t.ID)}
+	t.TaskInfo.TaskId = newTaskID(t.ID)
 	t.TaskInfo.SlaveId = details.GetSlaveId()
 	t.TaskInfo.Resources = []*mesos.Resource{
 		mesos.ScalarResource("cpus", containerCpus),
@@ -173,53 +150,9 @@ func newPodTask(ctx api.Context, pod *api.Pod, executor *mesos.ExecutorInfo) (*P
 	task := &PodTask{
 		ID:       taskId,
 		Pod:      pod,
-		TaskInfo: new(mesos.TaskInfo),
+		TaskInfo: newTaskInfo("PodTask"),
 		podKey:   key,
 	}
-	task.TaskInfo.Name = proto.String("PodTask")
 	task.TaskInfo.Executor = executor
 	return task, nil
-}
-
-/**
-HACK(jdef): we're not using etcd but k8s has implemented namespace support and
-we're going to try to honor that by namespacing pod keys. Hence, the following
-funcs that were stolen from:
-    https://github.com/GoogleCloudPlatform/kubernetes/blob/release-0.5/pkg/registry/etcd/etcd.go
-**/
-
-const PodPath = "/pods"
-
-// makeListKey constructs etcd paths to resource directories enforcing namespace rules
-func makeListKey(ctx api.Context, prefix string) string {
-	key := prefix
-	ns, ok := api.NamespaceFrom(ctx)
-	if ok && len(ns) > 0 {
-		key = key + "/" + ns
-	}
-	return key
-}
-
-// makeItemKey constructs etcd paths to a resource relative to prefix enforcing namespace rules.  If no namespace is on context, it errors.
-func makeItemKey(ctx api.Context, prefix string, id string) (string, error) {
-	key := makeListKey(ctx, prefix)
-	ns, ok := api.NamespaceFrom(ctx)
-	if !ok || len(ns) == 0 {
-		return "", fmt.Errorf("Invalid request.  Namespace parameter required.")
-	}
-	if len(id) == 0 {
-		return "", fmt.Errorf("Invalid request.  Id parameter required.")
-	}
-	key = key + "/" + id
-	return key, nil
-}
-
-// makePodListKey constructs etcd paths to pod directories enforcing namespace rules.
-func makePodListKey(ctx api.Context) string {
-	return makeListKey(ctx, PodPath)
-}
-
-// makePodKey constructs etcd paths to pod items enforcing namespace rules.
-func makePodKey(ctx api.Context, id string) (string, error) {
-	return makeItemKey(ctx, PodPath, id)
 }
